Update layout, make region selection dynamic
Updated the UI layout by wrapping text elements in Viewbox for better text scaling. The region selection has also been made dynamic, enabling the selection of the desired game server in the Welcome page. The selected regional routes are passed as parameters throughout the application for better integration with RiotGamesApi. Changed the default server from NA to EUW to reflect the more frequent usage. These changes provide more flexibility to the user and improve the usability of the application.