Refactor SetMatchTimeline method in MatchDetail.xaml.cs
Refactoring was done to make the code more readable and maintainable by eliminating code duplications. Instead of repeatedly declaring ellipse and rectangle properties for different instances within different conditional checks, they are now separately declared in new methods such as 'SetPrecedentEvent', 'SetPositiveEvent' and 'SetNegativeEvent', and invoked as and when required.