Refactor initializing team money collection
Removed the initial zero value from the ObservableCollection for team1Money and team2Money in the MatchDetail.xaml.cs. This change cleans up the instantiation of these collections, making the code cleaner and the interpretation more straightforward. Also, the README file has been moved from the NexusClient directory to the root directory.