Update account validation and asset paths
This commit updates the input validation in the welcome page to verify the summoner ID in a more granulated way. It also changes the paths of assets used in the application to point to the updated 14.1.1 version of data files instead of 13.24.1, and the Camille.RiotGames API. A fallback is implemented to retrieve MasterLeague data if there are fewer than three challengerEntries.