Refactor UI elements layout in MatchInfoPage and adjust text settings
In the MatchInfoPage.xaml, unnecessary Viewbox wrappers were removed to streamline layout. The layout in MatchInfoPage.xaml.cs was updated too: StackPanel used for showing match info was replaced with Grid for better column control. Modifications were also made to the font size and color arguments in setting the text, as well as to the image source. In addition, unnecessary stretch parameter was removed from the SetTextWithViewbox method in UtilisMethods.cs.